health/pyproject.toml
[project]
name = "health-agent"
version = "0.1.0"
description = "A Health Agent using Gemini"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]",
    "google-generativeai",
    "python-dotenv",
    "uvicorn",
    "requests",
]
